Weep in dispair as i try to understand GPU depth buffering, instance rendering, and translucency

+++ src/domain/VoidHuntersRevived.Domain.Pieces.Common/Components/Static/zIndex.cs
@@ -4,9 +4,9 @@ namespace VoidHuntersRevived.Domain.Pieces.Common.Components.Static
 {
     public struct zIndex : IEntityComponent
     {
-        public readonly int Value;
+        public readonly float Value;
 
-        public zIndex(int value)
+        public zIndex(float value)
         {
             Value = value;
         }
